requirements.txt
# Flask Core
Flask
python-dotenv

# Database
Flask-SQLAlchemy
Flask-Migrate

# Authentication & Forms
Flask-Login
Flask-WTF
Werkzeug
email_validator

# Data Science & ML
pandas
scikit-learn==1.5.1
joblib
geopy
requests